requirements/development.txt
-r base.txt
Faker==4.1.1
factory-boy==2.12.0
flake8==3.8.3
pytest-django==3.9.0
